requirements.txt
django==5.0.6
django_allauth==0.63.0
oauthlib==3.2.2
requests==2.31.0
requests-oauthlib==2.0.0
mysqlclient==2.2.4
python-dotenv==1.0.1
django-markdownify==0.9.5
markdown==3.6
daphne==4.1.2
channels==4.1.0
